# apigateway/openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce API Gateway
  version: 1.1.0
  description: Consolidated OpenAPI spec for the API Gateway proxying User, Product, and Order services.
servers:
  - url: http://localhost:8083
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
...
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer }
    CreateProductRequest:
      type: object
      required: [name, price, stock]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
    UpdateProductRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
    ReserveReleaseRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
  /api/v1/products/{productId}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Products]
      summary: Update product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProductRequest' }
      responses:
        '200': { description: Updated }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/products/{productId}/reserve:
    post:
      tags: [Products]
      summary: Reserve stock for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveReleaseRequest' }
      responses:
        '200':
          description: Reserved
          content:
            application/json:
              schema:
                type: object
                properties:
                  reserved: { type: integer }
                  stock: { type: integer }
        '409':
          description: Insufficient stock
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/products/{productId}/release:
    post:
      tags: [Products]
      summary: Release stock for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveReleaseRequest' }
      responses:
        '200':
          description: Released
          content:
            application/json:
              schema:
                type: object
                properties:
                  released: { type: integer }
                  stock: { type: integer }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/products/search:
    get:
      tags: [Products]
      summary: Search products
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number, format: float }
        - in: query
          name: maxPrice
          schema: { type: number, format: float }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }

  # Orders
...